Refactor complete.  Separate class for all Button event handlers and their necessary functions.
